refactory component abstract list
