+++ src/app/pages/categories/category-list/category-list.component.html
@@ -1,9 +1,4 @@
-<nav class="mb-5">
-  <ol class="breadcrumb">
-    <li class="breadcrumb-item"><a routerLink="/">Home</a></li>
-    <li class="breadcrumb-item active">Categorias</li>
-  </ol>
-</nav>
+<app-bread-crumb></app-bread-crumb>
 
 <div class="row mb-4">
   <div class="col-md">
@@ -25,7 +20,7 @@
       </thead>
 
       <tbody>
-        <tr *ngFor="let category of categories">
+        <tr *ngFor="let category of resources">
           <td>
             <strong>{{ category.name }}</strong>
             <br>
