+++ src/app/pages/categories/category-list/category-list.component.ts
@@ -1,35 +1,17 @@
 import { Component, OnInit } from '@angular/core';
 
 import { Category, CategoryService } from '../shared'
+import { BaseResourceListComponent } from 'src/app/shared/components/base-resource-list/base-resource-list.component';
 
 @Component({
   selector: 'app-category-list',
   templateUrl: './category-list.component.html',
   styleUrls: ['./category-list.component.css']
 })
-export class CategoryListComponent implements OnInit {
+export class CategoryListComponent extends BaseResourceListComponent<Category> {
 
-  categories: Category[] = []
+  constructor(private categoryService: CategoryService) {
 
-  constructor(private categoryService: CategoryService) { }
-
-  ngOnInit() {
-
-    this.categoryService
-            .findAll()
-            .subscribe(categories => this.categories = categories, error => alert('erro ao retornar lista'))
-  }
-
-  delete(category) {
-
-    let mustDelete = confirm('Deseja realmente excluir esse item ?')
-
-    if (mustDelete) {
- 
-      this.categoryService
-              .delete(category.id)
-              .subscribe(() => this.categories = this.categories.filter(element => element != category), 
-                        () => alert('erro ao tentar excluir'))
-    }
-  }
+    super(categoryService)
+   }
 }
