+++ src/app/pages/entries/entry-list/entry-list.component.html
@@ -1,9 +1,4 @@
-<nav class="mb-5">
-  <ol class="breadcrumb">
-    <li class="breadcrumb-item"><a routerLink="/">Home</a></li>
-    <li class="breadcrumb-item active">Lançamentos</li>
-  </ol>
-</nav>
+<app-bread-crumb></app-bread-crumb>
 
 <div class="row mb-4">
   <div class="col-md">
@@ -27,7 +22,7 @@
       </thead>
 
       <tbody>
-        <tr *ngFor="let entry of entries">
+        <tr *ngFor="let entry of resources">
           <td>
             <strong>{{ entry.name }}</strong><br>
             <small class="text-success">{{ entry.date }}</small><br>
